remove casts

+++ libduel.cpp
@@ -1998,18 +1998,18 @@ LUA_STATIC_FUNCTION(GetCurrentPhase) {
 }
 LUA_STATIC_FUNCTION(IsMainPhase) {
 	const auto phase = pduel->game_field->infos.phase;
-	lua_pushboolean(L, static_cast<int>(phase == PHASE_MAIN1 || phase == PHASE_MAIN2));
+	lua_pushboolean(L, (phase == PHASE_MAIN1 || phase == PHASE_MAIN2));
 	return 1;
 }
 LUA_STATIC_FUNCTION(IsBattlePhase) {
 	const auto phase = pduel->game_field->infos.phase;
-	lua_pushboolean(L, static_cast<int>(phase >= PHASE_BATTLE_START) && (phase <= PHASE_BATTLE));
+	lua_pushboolean(L, (phase >= PHASE_BATTLE_START) && (phase <= PHASE_BATTLE));
 	return 1;
 }
 LUA_STATIC_FUNCTION(IsPhase) {
 	check_param_count(L, 1);
 	const auto phase = lua_get<uint16_t>(L, 1);
-	lua_pushboolean(L, static_cast<int>(pduel->game_field->infos.phase == phase));
+	lua_pushboolean(L, pduel->game_field->infos.phase == phase);
 	return 1;
 }
 LUA_STATIC_FUNCTION(SkipPhase) {
